feat: Add recursive service
services:
  - Add service to abstract the behavior to make a recursive view call.
views:
  - Migrate all logic for recursive call into dedicated service class.
tests:
  - Update tests to use dedicated service class.
